refactor(evm): remove execute module
